Move CharStream into impl.javacc package
