+++ pmd-javascript/src/main/java/net/sourceforge/pmd/lang/ecmascript5/Ecmascript5TokenManager.java
@@ -7,7 +7,7 @@ package net.sourceforge.pmd.lang.ecmascript5;
 import java.io.Reader;
 
 import net.sourceforge.pmd.lang.TokenManager;
-import net.sourceforge.pmd.lang.ast.SimpleCharStream;
+import net.sourceforge.pmd.lang.ast.impl.javacc.SimpleCharStream;
 import net.sourceforge.pmd.lang.ecmascript5.ast.Ecmascript5ParserTokenManager;
 
 /**
