+++ pmd-matlab/etc/grammar/matlab.jj
@@ -21,7 +21,7 @@ options {
 PARSER_BEGIN(MatlabParser)
 package net.sourceforge.pmd.lang.matlab.ast;
 
-import net.sourceforge.pmd.lang.ast.CharStream;
+import net.sourceforge.pmd.lang.ast.impl.javacc.CharStream;
 import net.sourceforge.pmd.lang.ast.TokenMgrError;
 
 public class MatlabParser {
