+++ pmd-objectivec/etc/grammar/ObjC2.0.jj
@@ -21,7 +21,7 @@ package net.sourceforge.pmd.lang.objectivec.ast;
 import java.io.*;
 import java.util.*;
 
-import net.sourceforge.pmd.lang.ast.CharStream;
+import net.sourceforge.pmd.lang.ast.impl.javacc.CharStream;
 import net.sourceforge.pmd.lang.ast.TokenMgrError;
 
 /**
