+++ pmd-plsql/etc/grammar/PldocAST.jjt
@@ -155,7 +155,7 @@ package net.sourceforge.pmd.lang.plsql.ast;
 
 import java.io.*;
 import net.sourceforge.pmd.lang.ast.Node;
-import net.sourceforge.pmd.lang.ast.SimpleCharStream;
+import net.sourceforge.pmd.lang.ast.impl.javacc.SimpleCharStream;
 import net.sourceforge.pmd.lang.ast.TokenMgrError;
 
 public class PLSQLParser {
