+++ pmd-visualforce/etc/grammar/VfParser.jjt
@@ -14,7 +14,7 @@ options {
 PARSER_BEGIN(VfParser)
 package net.sourceforge.pmd.lang.vf.ast;
 
-import net.sourceforge.pmd.lang.ast.CharStream;
+import net.sourceforge.pmd.lang.ast.impl.javacc.CharStream;
 import net.sourceforge.pmd.lang.ast.TokenMgrError;
 
 public class VfParser {
