+++ pmd-vm/src/main/java/net/sourceforge/pmd/lang/vm/util/VelocityCharStream.java
@@ -1,7 +1,7 @@
 
 package net.sourceforge.pmd.lang.vm.util;
 
-import net.sourceforge.pmd.lang.ast.CharStream;
+import net.sourceforge.pmd.lang.ast.impl.javacc.CharStream;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
